+++ src/components/redux/slices/workerPanelState.ts
@@ -4,10 +4,12 @@ import { GetWorkerResponse } from "../../../types/stableHorde/api"
 interface WorkerPanelStateSliceType {
     sortKey: keyof GetWorkerResponse
     order: "asc" | "desc"
+    workerType: GetWorkerResponse["type"]
     workerFilter?: string
 }
 
 const initialState: WorkerPanelStateSliceType = {
+    workerType: "image",
     sortKey: "name",
     order: "asc"
 }
@@ -16,6 +18,9 @@ export const workerPanelStateSlice = createSlice({
     name: "workerPanel",
     initialState,
     reducers: {
+        setWorkerType: (state, action: PayloadAction<GetWorkerResponse["type"]>) => {
+            state.workerType = action.payload
+        },
         setWorkerFilter: (state, action: PayloadAction<string>) => {
             state.workerFilter = action.payload
         },
@@ -29,4 +34,4 @@ export const workerPanelStateSlice = createSlice({
 })
 
 // Action creators are generated for each case reducer function
-export const { setWorkerFilter, setSortKey, setOrder } = workerPanelStateSlice.actions
+export const { setWorkerType, setWorkerFilter, setSortKey, setOrder } = workerPanelStateSlice.actions