+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### Version 0.0.3 - 04/08/2023
+- Added Worker Type selector to Workers panel (image, interrogation, and text)
+- Fixed an issue caused by lexical sorting when sorting by performance key, now it is numerically parsed
+- Small service name refactor to reflect update from stableHorde to aiHorde
+
 ### Version 0.0.2 - 03/01/2023
 - Added Bridge Agent to worker card
 - Added Bridge Agent to worker list filtering, and as a Sort Key